requirements.txt
Flask==0.10.1
Flask-Admin==1.3
Flask-BabelEx==0.9.2
Flask-Config-Helper
Flask-Cors==2.1.0
Flask-DebugToolbar==0.10.0
Flask-Fixtures==0.3.3
Flask-Marshmallow==0.6.2
Flask-Migrate==1.4.0
Flask-OAuthlib==0.9.1
Flask-RESTful==0.3.3
Flask-Script==2.0.5
Flask-Security
Flask-SQLAlchemy==1.0
Flask-Swagger==0.2.12
Flask-WTF==0.11
flatdict==1.2.0
gunicorn==19.3.0
inflection==0.3.1
marshmallow-sqlalchemy==0.8.0
psycopg2==2.6.1
python-dateutil==2.4.2
pyyaml==3.11
shortuuid==0.4.2
SQLAlchemy==0.9.8
voluptuous==0.8.7
Werkzeug==0.10.4
-e git+https://github.com/yoophi/flask-swagger-ui.git@ecd065e667a902049a530bd0050acb7275896963#egg=Flask_Swagger_UI
